Added desciption how to build the library on Windows
Added support for glog >= 0.3.5

The version 0.3.5 of glog comes with support for CMake. The support
is implemented by config mode and not by a find script. As far as I know
the config mode does not support building the Object libraies. Maybe this
is a bug in the glog CMake config scripts or in CMake itself. A simple
workarround is to not use the object libraries in ifm3d.

Added windows shared library (DLL) support

It is not enough to declare CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS in CMake.
Like the CMake help states global data symbols need the
__declspec(dllimport) when compiling against this code in the .dll.

Added the include Path variables for xmlrpc-c

The xmlrpc-c find script provided by CMake is broken under windows.
This needs an upsdtream patch. Either in the xmlrpc-c repo by adding
a CMake config script. Or by fixing the upstream CMake find script.

Solution for the __attribute__((constructor)) under Windows

This is a solution borrowed from StackOverfolw.
https://stackoverflow.com/a/2390626/695442

Thanks to Joe Lowe for releasing this under the public domain.

Signed-off-by: Mateusz Marzec <mmarzec@jpembedded.eu>
Signed-off-by: Christian Ege <christian.ege@ifm.com>