Update list init for c++11 (For O3D303's poky toolchain)
The implementation of std::unordered_map in the stdlib in the O3D303
toolchain appears to not have been updated for C++14. As a result,
brace-initialization of an empty std::unordered_map selects an explicit
constructor, which is ill-formed.

The fix is, for the empty maps, enclose in an additional set of curly
braces to force copy-list initialization.

A detailed explanation of this issue can be found here:
https://stackoverflow.com/questions/26947704/implicit-conversion-failure-from-initializer-list