Invoking creators with empty sets for missing images
In the case where different chunks are present for a given frame (i.e.
unreliable transports like UDP, or changing of schema mid-stream) we
need to afford a mechanism to 'clear' images and prevent stale data from
propagating.

If a given chunk is not found, its format will be INVALID. In this case,
invoke the appropriate ImCreate/CloudCreate functions with zero-sized
data to signal data should be cleared.