Add VS Code devlopment within a Docker Container
The dependencies for the ifm3d might become tricky for some
environments. For me it is convienient to perform the development inside
a defined Docker container. All required dev packages are installed
inside the Container and Visual Studio Code can run the builds and the
terminal in this context.

As a bonus there comes a shell script for installing all the dependencies.

Signed-off-by: Christian Ege <christian.ege@ifm.com>